Add test case for principal zero principal amount

+++ loan-calculator-api/LoanCalculator/Core/LoanCalculator.cs
@@ -4,7 +4,7 @@
     {
         public double PrincipalLoanAmount { get; set; }
         public double InterestRate { get; set; }
-        public double NumberOfPayments { get; set; }
+        public int NumberOfPayments { get; set; }
         public abstract double CalculateLoanInstallment();
     }
 }
